feat(cms): add upgrade preview page
